Separate config and rollout; introduce ausroller module
* This commit separates the code for configuration (reading cli
  parameter, reading config file) and the actual code to rollout.
