+++ ausroller/__init__.py
@@ -1,0 +1,2 @@
+from config import Configuration
+from core import Ausroller
